week_4/requirements.txt
google-generativeai>=0.3.2
Pillow>=10.0.0
numpy>=1.24.0
mcp>=0.1.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
opencv-python>=4.8.0
tkinter  # Usually comes with Python standard library 